Moved utils.js into lib to be used also in tests
